CAMS-37: WIP - fetch.post wrapper, temporary test disable
Wrapped fetch POST method in a the http utility module, in order to make
it easy to mock fetch requests.  Eventually we'll want to add GET, PUT,
DELETE, etc.

Temporarily disabled test in lib/use-cases/chapter-15-case-list.test.ts.
Will need to revisit.
